fixed height/clear related bugs, rewrote grouping
Changed some height calculations to use this.element.dom.clientHeight,
as this.getHeight() wasn't always returning a non-zero answer.
Fixed bug where calling store.removeAll() or removing the last item in a
store would cause the list to be rendered incorrectly.
Fixed showEmptyText.
Rewrote the grouping code to address some problems.  First, the way the
grouping was working you had to group things by the same grouping as the
letters on the Index bar.  My rewrite makes a separation between the
mapping for the list groups and the mapping for the index bar.  It also
doesn't convert to lowercase but instead uses uppercase like the
indexbar does already.
I also made it where when you jump to a letter that's not represented in
the data, it jumps to the last record in the previous group.  It
basically jumps to where you WOULD be, if it had a list item with the
letter you tapped.