uppercased group ids
Uppercased the group ids in the groupMap.  Otherwise, I found that
problems could be caused when the same group key existed and the data
was sorted in a case-insensitive fashion.  Perhaps this should be
configurable?  Probably not, though, as indexBars don't typically have
separate items for upper- and lower-case.  As such, I'm not sure there's
much of a common use case other than doing it this way.